refactor: Omission field file structure
moved omission field to omission field array component
