+++ src/components/RulesetFormSections/OmissionFieldArray/OmissionField/OmissionField.test.js
@@ -3,8 +3,8 @@ import { renderWithIntl, TestForm, Select } from '@folio/stripes-erm-testing';
 
 import OmissionField from './OmissionField';
 
-import { translationsProperties } from '../../../../test/helpers';
-import { refdata as mockRefdata } from '../../../../test/resources';
+import { translationsProperties } from '../../../../../test/helpers';
+import { refdata as mockRefdata } from '../../../../../test/resources';
 
 const omission = {
   timeUnit: {
@@ -22,8 +22,8 @@ const omission = {
   },
 };
 
-jest.mock('../../utils', () => ({
-  ...jest.requireActual('../../utils'),
+jest.mock('../../../utils', () => ({
+  ...jest.requireActual('../../../utils'),
   useSerialsManagementRefdata: () => mockRefdata,
 }));
