+++ src/components/RulesetFormSections/index.js
@@ -3,7 +3,6 @@ export { default as ModelRulesetSelection } from './ModelRulesetSelection';
 export { default as PatternTimePeriodForm } from './PatternTimePeriodForm';
 export { default as IssuePublicationFieldArray } from './IssuePublicationFieldArray';
 export { default as IssuePublicationField } from './IssuePublicationField';
-export { default as OmissionField } from './OmissionField';
 export { default as OmissionFieldArray } from './OmissionFieldArray';
 export { default as CombinationFieldArray } from './CombinationFieldArray';
 export { default as ChronologyFieldArray } from './ChronologyFieldArray';
